neobistime/requirements.txt
django==2.2.13
djangorestframework==3.10.3
python-decouple==3.1
psycopg2==2.8.4
uvicorn==0.11.2
gunicorn==20.0.4
django-allauth==0.42.0
drf-yasg==1.17.1
django-rest-auth==0.9.5
flake8==3.8.3
celery==4.4.6
